Update LootPop.lua
Made improvements to /lootpop config display and functionalities. Changing scale, spacing or frame strata while the anchor is displaying will now show these changes on the anchor without having to close and reopen the anchor. Added an option in the config to have the popups grow up or grow down so you can position the anchor closer to the top or bottom of the screen and control the direction of the popups so they aren't going off the screen. Fixed an issue with the position of the display moving to a different spot after logging out and only reverted to the intended spot via opening /lootpop config and click the Show/Move Anchor button. Fixed an issue that caused the length of the bar to not appropriately change if the name of the item was very short.